pyproject.toml
[project]
name = "fastapi-di-demo"
version = "0.1.0"
description = "A injection-based FastAPI demo application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "uvicorn",
    "pydantic",
    "dependency-injector>=4.47.1",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pyrefly"
]
[tool.pyrefly]
project_includes = ["**/*"]
project_excludes = [
    "**/.[!/.]*",
    "**/*venv/**/*",
]
